+++ accessories/light.js
@@ -83,9 +83,15 @@ class LightAccessory extends SwitchAccessory {
       // Find hue closest to the one requested
       const foundValues = this.dataKeys('hue');
       const closest = foundValues.reduce((prev, curr) => Math.abs(curr - state.hue) < Math.abs(prev - state.hue) ? curr : prev);
-      const hexData = data[`hue${closest}`];
-
-      log(`${name} setHue: (closest: hue${closest})`);
+      var hexData = "";
+      // If saturation is less than 10, choose white
+      if (state.saturation < 10 && data[`white`]) {
+        hexData = data[`white`];
+        log(`${name} setHue: (closest: white)`);
+      } else {
+        hexData = data[`hue${closest}`];
+        log(`${name} setHue: (closest: hue${closest})`);
+      }
       await this.performSend(hexData);
     });
   }