+++ CHANGELOG.md
@@ -4,9 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.4.7] - 2021-11-26
+### Added
+ - Adds 0x6184 device support
+ - Added support for 'white' light. Thanks to @JuniorGenius
+### Fixed
+ - Removed unsupported AutoOnOff code from heater-cooler accessory which was causing exceptions.
+
 ## [4.4.6] - 2021-08-03
 ### Added
- - -Adds support for 0x6508 devices
+ - Adds support for 0x6508 devices
 ### Changed
  - General code tidy-up/standardisation.
 ### Fixed